requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
Deprecated==1.2.13
Django==3.2.13
djangorestframework==3.13.1
flake8==4.0.1
flake8-builtins==1.5.3
flake8-docstrings==1.6.0
flake8-import-order==0.18.1
GPUtil==1.4.0
importlib-metadata==4.2.0
mccabe==0.6.1
packaging==21.3
psutil==5.9.1
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.9
pytz==2022.1
redis==4.3.3
sqlparse==0.4.2
typing-extensions==4.2.0
wrapt==1.14.1
zipp==3.8.0
